Adjust padding in ToDo option TextField.
