+++ app/src/main/java/com/example/todo_app/ui/feature/toDoOptions/ToDoOptions.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -219,7 +218,7 @@ private fun TextFieldOption(
                 Box(
                     contentAlignment = contentAlign,
                     modifier = modifier
-                        .padding(8.dp)
+                        .padding(10.dp)
                 ) {
                     if (textState.value.isEmpty() && !focusState.value) {
                         Text(
